[QML] Table view param column width int -> double
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@12299 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtrest/CTableViewParamRepresentationControllerComp.cpp
@@ -106,7 +106,7 @@ bool CTableViewParamRepresentationControllerComp::GetDataModelFromRepresentation
 		}
 
 		if (representation.ContainsKey("Size", i)){
-			int size = representation.GetData("Size", i).toInt();
+			double size = representation.GetData("Size", i).toDouble();
 			headerInfo.size = size;
 		}
